Add testing
This commit adds some basic testing with pytest. It also sets up an
environment with Poetry and runs the tests automatically using GitHub
actions.

The file pre_pilot_script.py was renamed to report.py and was refactored
a little bit in order to make it easier to test. The Python files were
also reformatted with black.