Replace poetry with venv
The latest released version of whisper isn't installable from PyPI with Poetry. But the latest commit on GitHub
is. In order to stick with the latest released version this switches to using a virtualenv for managing dependencies. This also removes a dependency, and having to learn a new tool when it isn't really necessary.