Adding diffs and run ids
This commit adds the generation of HTML diffs using Python's difflib, which closes #10.

Also each run is assigned a `run_id` that appears in the report csv and which is used to name the output files. Closes #9

Finally the code uses an explicit manifest of files, and their
transcripts to test in the data.csv which takes some of the guesswork
out of which files are being tested. Fixes #11